+++ test/unit/handlers/inbound-api.test.ts
@@ -47,7 +47,7 @@ import {
   OutboundAccountsModel
 } from '~/models/outbound/accounts.model'
 import { PISPTransactionPhase } from '~/models/pispTransaction.interface'
-import { notificationChannel } from '~/models/pispConsentRequest.model'
+import { PISPOTPValidateModel } from '~/models/outbound/pispOTPValidate.model'
 import ThirdpartyAuthorizations from '~/handlers/inbound/thirdpartyRequests/transactions/{ID}/authorizations'
 import ConsentsHandler from '~/handlers/inbound/consents'
 import ConsentRequestsIdHandler from '~/handlers/inbound/consentRequests/{ID}'
@@ -681,7 +681,7 @@ describe('Inbound API routes', (): void => {
       expect(result.statusCode).toEqual(202)
       expect(toolkit.getPubSub).toBeCalledTimes(1)
 
-      const channel = notificationChannel(postConsentRequest.consentRequestId)
+      const channel = PISPOTPValidateModel.notificationChannel(postConsentRequest.consentRequestId)
       expect(pubSubMock.publish).toBeCalledWith(channel, postConsentRequest)
     })